@@ -22,10 +22,18 @@ buildscript {
         maven { url "https://www.jitpack.io" }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
+        classpath 'com.android.tools.build:gradle:3.2.0-beta03'
         classpath 'io.fabric.tools:gradle:1.+'
     }
 }
+
+allprojects {
+    repositories {
+        google()
+        jcenter()
+    }
+}
+
 apply plugin: 'com.android.application'
 if (rootProject.file('private.properties').exists()) {
     apply plugin: 'io.fabric'
@@ -54,16 +62,16 @@ def getRevision = { ->
 }
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.1'
     testBuildType 'debugTest'
-    useLibrary 'org.apache.http.legacy'
 
     defaultConfig {
         versionCode 195
         versionName "4.10"
         minSdkVersion 14
-        targetSdkVersion 27
+        targetSdkVersion 28
+        multiDexEnabled true
         applicationId "com.irccloud.android"
         testApplicationId "com.irccloud.android.test"
         testInstrumentationRunner "android.test.InstrumentationTestRunner"
@@ -329,26 +337,22 @@ tasks.whenTaskAdded { task ->
 }
 
 ext {
-    supportLibVersion = '27.1.1'
-    playServicesVersion = '12.0.1'
     dbflowVersion = "4.2.4"
 }
 
 dependencies {
-    implementation "com.android.support:support-v4:$supportLibVersion"
-    implementation "com.android.support:support-v13:$supportLibVersion"
-    implementation "com.android.support:appcompat-v7:$supportLibVersion"
-    implementation "com.android.support:cardview-v7:$supportLibVersion"
-    implementation "com.android.support:design:$supportLibVersion"
-    implementation "com.android.support:recyclerview-v7:$supportLibVersion"
-    implementation "com.android.support:customtabs:$supportLibVersion"
-    implementation "com.android.support:support-emoji:$supportLibVersion"
-    implementation "com.android.support:exifinterface:$supportLibVersion"
-    implementation "com.google.android.gms:play-services-base:$playServicesVersion"
-    implementation "com.google.android.gms:play-services-auth:$playServicesVersion"
-    implementation "com.google.android.gms:play-services-identity:$playServicesVersion"
-    implementation "com.google.android.gms:play-services-ads:$playServicesVersion"
-    implementation "com.google.android.gms:play-services-gcm:$playServicesVersion"
+    implementation "androidx.appcompat:appcompat:1.0.0-beta01"
+    implementation "androidx.cardview:cardview:1.0.0-beta01"
+    implementation "com.google.android.material:material:1.0.0-beta01"
+    implementation "androidx.recyclerview:recyclerview:1.0.0-beta01"
+    implementation "androidx.browser:browser:1.0.0-beta01"
+    implementation "androidx.emoji:emoji:1.0.0-beta01"
+    implementation "androidx.exifinterface:exifinterface:1.0.0-beta01"
+    implementation "com.google.android.gms:play-services-base:15.0.1"
+    implementation "com.google.android.gms:play-services-auth:15.0.1"
+    implementation "com.google.android.gms:play-services-identity:15.0.1"
+    implementation "com.google.android.gms:play-services-ads:15.0.1"
+    implementation "com.google.android.gms:play-services-gcm:15.0.1"
     implementation 'com.google.android.apps.dashclock:dashclock-api:+'
     implementation 'com.vandalsoftware.android:dslv:+'
     implementation 'com.damnhandy:handy-uri-templates:2.0.2'
@@ -363,6 +367,7 @@ dependencies {
     implementation 'org.solovyev.android.views:linear-layout-manager:0.5@aar'
     implementation 'com.datatheorem.android.trustkit:trustkit:1.0.0'
     implementation 'com.theartofdev.edmodo:android-image-cropper:2.6.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.10.0'
     implementation('com.crashlytics.sdk.android:crashlytics:2.9.1@aar') {
         transitive = true
     }
