@@ -19,7 +19,7 @@ package com.irccloud.android;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 public class CrashlyticsLeakService extends Service {
 
