@@ -20,16 +20,16 @@ import android.content.Context;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.text.emoji.EmojiCompat;
-import android.support.text.emoji.widget.EmojiEditTextHelper;
-import android.support.v13.view.inputmethod.InputConnectionCompat;
-import android.support.v13.view.inputmethod.InputContentInfoCompat;
-import android.support.v4.widget.DrawerLayout;
+import androidx.emoji.text.EmojiCompat;
+import androidx.emoji.widget.EmojiEditTextHelper;
+import androidx.core.view.inputmethod.InputConnectionCompat;
+import androidx.core.view.inputmethod.InputContentInfoCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
 import android.util.AttributeSet;
 import android.view.KeyEvent;
 import android.view.inputmethod.EditorInfo;
 import android.view.inputmethod.InputConnection;
-import android.support.v13.view.inputmethod.EditorInfoCompat;
+import androidx.core.view.inputmethod.EditorInfoCompat;
 
 // An EditText that lets you use actions ("Done", "Go", etc.) on multi-line edits.
 // From: http://stackoverflow.com/a/12570003/1406639
