@@ -21,8 +21,8 @@ import android.graphics.Color;
 import android.graphics.Typeface;
 import android.os.Build;
 import android.preference.PreferenceManager;
-import android.support.text.emoji.EmojiCompat;
-import android.support.v4.content.res.ResourcesCompat;
+import androidx.emoji.text.EmojiCompat;
+import androidx.core.content.res.ResourcesCompat;
 import android.text.Editable;
 import android.text.Html;
 import android.text.Spannable;
