@@ -16,8 +16,8 @@
 
 package com.irccloud.android.data;
 
-import android.databinding.BindingAdapter;
-import android.support.v4.content.ContextCompat;
+import androidx.databinding.BindingAdapter;
+import androidx.core.content.ContextCompat;
 import android.view.View;
 import android.view.ViewGroup;
 
