@@ -1,6 +1,6 @@
 package com.irccloud.android.data;
 
-import android.databinding.BindingAdapter;
+import androidx.databinding.BindingAdapter;
 import android.view.View;
 import android.view.ViewGroup;
 
